Move venues, screen groups and screens entities in core
