+++ web/src/app/content-management/venues/edit-tree-view-node-form/edit-tree-view-node-form.component.ts
@@ -4,7 +4,7 @@ import {Content} from "../../../content/content";
 import {NotificationService} from "../../../shared/notifications/notification.service";
 
 import * as _ from 'lodash';
-import {Venue} from "../entities/venue";
+import {Venue} from "../../../core/entities/venue";
 import {Setting} from "../../../settings/entities/setting";
 import {SettingStateHolderService} from "../../../settings/setting-state-manager/settings-state-holder.service";
 import {VenuesTreeViewService} from "../venues-tree-view/venues-tree-view.service";
