+++ web/src/app/content-management/venues/venues.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { environment } from '../../../environments/environment';
-import {Venue} from "./entities/venue";
+import {Venue} from "../../core/entities/venue";
 import { Observable } from 'rxjs/Observable';
 import {Content} from "../../content/content";
 import {ContentService} from "../../content/content.service";
