+++ web/src/app/maintenance/entities/venue-maintenance-info.ts
@@ -1,4 +1,4 @@
-import {Venue} from "../../content-management/venues/entities/venue";
+import {Venue} from "../../core/entities/venue";
 import {AutoupdateSchedule} from "./autoupdate-schedule";
 
 export class VenueMaintenanceInfo extends Venue {
