+++ web/src/app/maintenance/maintenance.service.ts
@@ -9,7 +9,7 @@ import {KioskVersionService} from "./kiosk-version.service";
 import {MaintenanceProperties} from "./entities/maintenance-properties";
 import {AutoupdateSchedule} from "./entities/autoupdate-schedule";
 import {KioskVersionDetails, KioskVersionDetailsMap} from "./entities/kiosk-version-details";
-import {Venue} from "../content-management/venues/entities/venue";
+import {Venue} from "../core/entities/venue";
 
 @Injectable()
 export class MaintenanceService {
